image_to_text/requirements.txt
pytesseract
numpy 
Flask
python-dotenv
Pillow